refactor(parsercore): split owned lexer snapshot test helper
- Move the owned lexer snapshot helper into its own test file to keep lifecycle tests focused.
- Use the recovery lineage fork table in the recovery test to match the shared-state scenario.

Buckley-Change-Hash: sha256:44af80554d5e75faa87d8ab594f2fb03ebf8ddd4538ae6060f85e7d2e11961c4
Buckley-Change-Stats: files=3 insertions=32 deletions=27 binaries=0
