+++ parsercore_phase0_per_version_lexer_lifecycle_internal_test.go
@@ -25,32 +25,6 @@ func (diagnosticParserCoreOwnedLexerPanicRestoreScanner) Scan(any, *ExternalLexe
 	return false
 }
 
-func newDiagnosticParserCoreOwnedLexerSnapshot(
-	t *testing.T,
-	compact *core.Core,
-	language *Language,
-	lexerPosition int,
-) *diagnosticParserCoreVersionLexerSnapshot {
-	t.Helper()
-	var snapshot *diagnosticParserCoreVersionLexerSnapshot
-	err := compact.ApplySchedulerAtomic(func(owner core.SchedulerTransactionToken) error {
-		var snapshotErr error
-		snapshot, snapshotErr = newDiagnosticParserCoreVersionLexerSnapshot(
-			compact,
-			language,
-			owner,
-			dfaRelexSnapshot{lexerPos: lexerPosition},
-			0,
-			0,
-		)
-		return snapshotErr
-	})
-	if err != nil {
-		t.Fatalf("construct owned lexer snapshot: %v", err)
-	}
-	return snapshot
-}
-
 func newDiagnosticParserCoreOwnedLexerRequest(
 	electionIndex int,
 	state StateID,
